Update and rename x Triangle.py to xTriangle -extra credit/1st attempt.py
Create a program that prints out a triangle of X's. Such as...

X
XX
XXX
XXXX
XXXXX

The "height" of the triangle (5 in the above case) will be enter first by the user and then you will display the triangle.

Use a loop to complete this. Try and use the fewest lines of code for this program.